+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - add missing json's for Lumium, Signalum & Enderium
 - add Lumium Textures
 - fix incorrect lang entries for ore veins
+- add Native Iridium Ore, Textures and Vein